requirements.txt
pandas~=2.3.3
numpy~=2.3.4
psycopg2-binary~=2.9.11
boto3~=1.40.55
fastapi~=0.119.0
uvicorn~=0.29.0
duckdb~=1.4.1
ta~=0.11.0
sqlalchemy~=2.0.44
pyspark~=4.0.1
python-multipart~=0.0.9
pyarrow~=16.1.0
fastparquet~=2024.2.0
